+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Added `tidy3d.plugins.metrics` module for constructing and serializing simulation metrics like `ModeCoefficient` and `ModePower`.
+- Added `tidy3d.plugins.expressions` module for constructing and serializing mathematical expressions and simulation metrics like `ModeAmps` and `ModePower`.
 
 ## [2.7.3] - 2024-09-12